working puppetdb ebuild
This is based on the work in https://github.com/optiz0r/gentoo-overlay/. It
depends on the leiningen package from there. Most of it is probably a
rather dirty hack and the install is also rather binary.
Finally I have a way of installing a modern puppet setup on my personal
dev cloud.